Update image count performance and add deployment workflow
