+++ .github/workflows/deploy.prod.yml
@@ -1,12 +1,12 @@
-name: Run Foodish images check
+name: Deploy to Production
 
 on:
-  pull_request:
+  push:
     branches: [ main ]
 
 jobs:
-  build:
-    name: Build and Test
+  deploy:
+    name: Deploy
     runs-on: ubuntu-latest
 
     steps:
@@ -17,5 +17,5 @@ jobs:
           node-version: '16.x'
       - name: Install dependencies
         run: npm ci
-      - name: Run image checks
-        run: npm test
+      - name: Run deployment scripts
+        run: npm run deploy-script
