+++ scripts/deployment/imageMetaData.json
@@ -1,0 +1,13 @@
+{
+  "biryani": 81,
+  "burger": 87,
+  "butter-chicken": 22,
+  "dessert": 36,
+  "dosa": 83,
+  "idly": 77,
+  "pasta": 34,
+  "pizza": 95,
+  "rice": 35,
+  "samosa": 22,
+  "total": 572
+}
